# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "木棉手札2.0的api詳細文件 \n\n**請注意本文`status code`為自訂放在`json`中  而不是 `http status code`**\n",
        "version": "1.0.0",
        "title": "木棉手札2.0 API"
    },
    "host": "dev.dacsc.club",
    "schemes": [
        "https"
    ],
    "basePath": "/v1",
    "paths": {
...
        "/attitudestatus": {
            "get": {
                "tags": [
                    "attitudestatus"
                ],
                "description": "得到獎懲紀錄",
                "responses": {
                    "103": {
                        "description": "token過期",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/AttitudeStatusResponse"
                        }
                    },
                    "400": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Token_Auth": []
                    }
                ]
            }
        },
        "/absentstate": {
            "get": {
                "tags": [
                    "absentstate"
                ],
                "description": "得到出勤紀錄",
                "responses": {
                    "103": {
                        "description": "token過期",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AbsentstateResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Token_Auth": []
                    }
                ]
            }
        },
        "/calendar": {
            "get": {
                "tags": [
                    "calendar"
                ],
                "description": "得到日曆",
                "responses": {
                    "103": {
                        "description": "token過期",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CalenderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Token_Auth": []
                    }
                ]
            }
        },
        "/curriculum": {
            "get": {
                "tags": [
                    "curriculum"
                ],
                "description": "查詢課表",
                "responses": {
                    "103": {
                        "description": "token過期",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "week1": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/TimeTableResponse"
                                    }
                                },
                                "week2": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/TimeTableResponse"
                                    }
                                },
                                "week3": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/TimeTableResponse"
                                    }
                                },
                                "week4": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/TimeTableResponse"
                                    }
                                },
                                "week5": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/TimeTableResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Token_Auth": []
                    }
                ]
            }
        },
        "/QandA": {
            "get": {
                "tags": [
                    "QandA"
                ],
                "description": "查詢學校",
                "responses": {
                    "103": {
                        "description": "token過期",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/QandAResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "Token_Auth": []
                    }
                ]
            }
        }
    },
    "securityDefinitions": {
        "Token_Auth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "description": "錯誤訊息"
                },
                "code": {
                    "type": "number",
                    "description": "錯誤代碼"
                }
            }
        },
...
        "AttitudeStatusResponse": {
            "type": "object",
            "required": [
                "status",
                "count"
            ],
            "properties": {
                "status": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Attitude"
                    }
                },
                "count": {
                    "type": "object",
                    "required": [
                        "smallcite",
                        "smallfault",
                        "middlecite",
                        "middlefault",
                        "bigcite",
                        "bigfault"
                    ],
                    "properties": {
                        "smallcite": {
                            "type": "number"
                        },
                        "smallfault": {
                            "type": "number"
                        },
                        "middlecite": {
                            "type": "number"
                        },
                        "middlefault": {
                            "type": "number"
                        },
                        "bigcite": {
                            "type": "number"
                        },
                        "bigfault": {
                            "type": "number"
                        }
                    }
                }
            }
        },
        "Attitude": {
            "type": "object",
            "required": [
                "date",
                "item",
                "text"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "item": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "AbsentstateResponse": {
            "type": "object",
            "required": [
                "date",
                "type",
                "class"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "type": {
                    "type": "string"
                },
                "class": {
                    "type": "string"
                }
            }
        },
        "CalenderResponse": {
            "type": "object",
            "required": [
                "date",
                "content"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "TimeTableResponse": {
            "type": "object",
            "required": [
                "start",
                "end",
                "subject",
                "teacher"
            ],
            "properties": {
                "start": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "teacher": {
                    "type": "string"
                }
            }
        },
        "QandAResponse": {
            "type": "object",
            "required": [
                "question",
                "answear"
            ],
            "properties": {
                "question": {
                    "type": "string"
                },
                "answear": {
                    "type": "string"
                }
            }
        }
    }
}